requirements.txt
discord
#pyautogui
pyzmq
flask
APScheduler==3.9.1
asgiref==3.5.2
#backports.zoneinfo
certifi==2022.6.15
charset-normalizer==2.1.0
coloredlogs==15.0.1
#ddddocr==1.4.3
Django==4.0.6
django-apscheduler==0.6.2
django-bootstrap3==21.2
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-tables2==2.4.1
django-environ==0.9.0
djangorestframework==3.13.1
flatbuffers==2.0
humanfriendly==10.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mpmath==1.2.1
mysqlclient==2.1.1
numpy==1.22.4
#onnxruntime==1.12.0
opencv-python==4.7.0.68
opencv-python-headless==4.7.0.68
packaging==21.3
Pillow==9.2.0
protobuf==4.21.3
pyparsing==3.0.9
pyreadline3==3.4.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
sympy==1.10.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
pbr==5.5.1
#sseclient-py
pandas
pyppeteer
pyppeteer_stealth
nest_asyncio
# pyvirtualdisplay
EasyProcess
aircv
openai
flask_cors
python-engineio==3.13.2
python-socketio==4.6.0
eventlet==0.33.3
pyperclip
azure.cognitiveservices.speech
langid
tenacity
rembg